% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ins_outs.R
\name{get_in_out_step}
\alias{get_in_out_step}
\title{gets inputs and outputs of code}
\usage{
get_in_out_step(sc, step)
}
\arguments{
\item{sc}{source code}

\item{step}{which step to get from}
}
\description{
gets inputs and outputs of code
}
