% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subcode.R
\name{get_subcode}
\alias{get_subcode}
\title{get a sub-section of the code}
\usage{
get_subcode(raw_code, start_line, end_line, inputs, outputs)
}
\arguments{
\item{raw_code}{lines of code}
}
\description{
get a sub-section of the code
}
